% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R, R/utils.R
\name{add_areasqkm}
\alias{add_areasqkm}
\title{Compute km2 area}
\usage{
add_areasqkm(x)

add_areasqkm(x)
}
\arguments{
\item{x}{An `sf` object with polygonal geometry. If `x` is not in a projected
CRS, `sf::st_area()` will compute ellipsoidal areas when possible.}
}
\value{
numeric

A numeric vector of areas in square kilometers.
}
\description{
Safely compute area in square kilometers (numeric vector)

Safely compute polygon area in km^2 and return a plain numeric vector.
}
\examples{
\dontrun{
library(sf)
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
add_areasqkm(nc[1, ])
}

}
