% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_properties.R
\name{add_hydroseq}
\alias{add_hydroseq}
\title{Compute and add the hydrosequence to a directed acyclic network.}
\usage{
add_hydroseq(
  topology,
  id = "flowpath_id",
  toid = "flowpath_toid",
  colname = "hydroseq"
)
}
\arguments{
\item{topology}{A data frame (or tibble) containing at least the identifier column
given by `id` and the downstream pointer column given by `toid`.}

\item{id}{Character scalar. Column name in `topology` with unique node identifiers.
Defaults to `"flowpath_id"`.}

\item{toid}{Character scalar. Column name in `topology` with the *downstream* node
identifier for each row. Use `NA` or `0` for outlets/terminals.
Defaults to `"flowpath_toid"`.}

\item{colname}{Character scalar. Column name to use in result.
Defaults to `"hydroseq"`}
}
\value{
The data frame `topology` with an additional column, named `colname`, representing the hydrosequence.
}
\description{
Compute and add the hydrosequence to a directed acyclic network.
}
