% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lengthkm}
\alias{add_lengthkm}
\title{Compute length in kilometers (numeric)}
\usage{
add_lengthkm(x)
}
\arguments{
\item{x}{An `sf` object with LINE* geometry. If `x` is not in a projected
CRS, `sf::st_length()` will compute ellipsoidal areas when possible.}
}
\value{
A numeric vector of lengths in kilometers.
}
\description{
Safely compute linestring length in km and return a plain numeric vector.
}
