% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_ogr.R
\name{as_ogr}
\alias{as_ogr}
\alias{as_ogr.character}
\alias{as_ogr.OGRSQLConnection}
\title{Delayed read for vector resources}
\usage{
as_ogr(x, layer, ..., query = NA, ignore_lyrs = .hf_ignore_lyrs)

\method{as_ogr}{character}(x, layer, ..., query = NA, ignore_lyrs = .hf_ignore_lyrs)

\method{as_ogr}{OGRSQLConnection}(x, layer, ..., query = NA, ignore_lyrs = .hf_ignore_lyrs)
}
\arguments{
\item{x}{the data source (file path, url, or database connection)}

\item{layer}{layer name (varies by driver, may be a file name without
extension); in case \code{layer} is missing, \code{st_read} will read the
first layer of \code{dsn}, give a warning and (unless \code{quiet = TRUE})
print a message when there are multiple layers, or give an error if there
are no layers in \code{dsn}. If \code{dsn} is a database connection, then
\code{layer} can be a table name or a database identifier (see
\code{\link[DBI]{Id}}). It is also possible to omit \code{layer} and rather
use the \code{query} argument.}

\item{...}{parameter(s) passed on to \link[sf]{st_as_sf}}

\item{query}{SQL query to pass in directly}

\item{ignore_lyrs}{Regular expression matching tables that are data-source
plumbing rather than user layers, excluded when `layer` is not given. The
default drops GeoPackage spec tables (`^gpkg_`), spatial indexes
(`^rtree_`), SQLite internals (`^sqlite_`), and the two tables QGIS writes
into a GeoPackage when you save a style or a project to it
(`layer_styles`, `qgis_projects`). Patterns are anchored, so a layer whose
name merely contains one of these is not dropped.}
}
\value{
a 'tbl_OGRSQLConnection'
}
\description{
A lazy data frame for GDAL vector data sources. as_ogr is DBI compatible and designed to work with dplyr.
}
\details{
The output of `as_ogr()` is a 'tbl_OGRSQLConnection` that extends `tbl_dbi` and
may be used with functions and workflows in the normal DBI way, see [OGRSQL()] for
the as_ogr DBI support.

To obtain an in memory data frame use an explicit `collect()` or `st_as_sf()`.
A call to `collect()` is triggered by `st_as_sf()` and will add the sf class
to the output.
}
\examples{
\dontrun{
library(dplyr)
tbl <- as_ogr("hydrofabric.gpkg", "flowpaths")
tbl |> filter(order >= 4) |> st_as_sf()
}
}
