% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flowpaths_to_linestrings}
\alias{flowpaths_to_linestrings}
\title{Convert MULTILINESTRINGS to LINESTRINGS}
\usage{
flowpaths_to_linestrings(flowpaths)
}
\arguments{
\item{flowpaths}{a flowpath `sf` object}
}
\value{
a `sf` object
}
\description{
Convert MULTILINESTRINGS to LINESTRINGS
}
\examples{
\dontrun{
fl <- sf::read_sf("hydrofabric.gpkg", "flowpaths")
fl <- flowpaths_to_linestrings(fl)
}
}
