% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_node}
\alias{get_node}
\title{Get endpoint or startpoint of LINESTRING}
\usage{
get_node(x, position = "end")
}
\arguments{
\item{x}{sf LINESTRING}

\item{position}{"start" or "end"}
}
\value{
sf POINT
}
\description{
Get endpoint or startpoint of LINESTRING
}
