% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{layer_exists}
\alias{layer_exists}
\title{Check if a layer exists in a geopackage}
\usage{
layer_exists(gpkg, name)
}
\arguments{
\item{gpkg}{path to .gpkg}

\item{name}{layer name}
}
\value{
logical
}
\description{
Check if a layer exists in a geopackage
}
