% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lynker_spatial_client}
\alias{lynker_spatial_client}
\title{Get an OAuth2 Client for Lynker Spatial}
\usage{
lynker_spatial_client()
}
\description{
Get an OAuth2 Client for Lynker Spatial
}
\keyword{internal}
