% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lynker_spatial_token}
\alias{lynker_spatial_token}
\title{Get an OAuth2 token for Lynker Spatial}
\usage{
lynker_spatial_token(..., client = lynker_spatial_client())
}
\arguments{
\item{...}{Unused; reserved for future options.}

\item{client}{An OIDC client, as returned by [lynker_spatial_client()].}
}
\value{
An httr2 OAuth2 token.
}
\description{
Get an OAuth2 token for Lynker Spatial
}
\examples{
\dontrun{
token <- lynker_spatial_token()
}
}
