% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lynker_spatial_url}
\alias{lynker_spatial_url}
\title{Create an authenticated URL connection for Lynker Spatial}
\usage{
lynker_spatial_url(url)
}
\arguments{
\item{url}{URL passed to the connection}
}
\value{
a URL connection object
}
\description{
Create an authenticated URL connection for Lynker Spatial
}
