% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{network_is_dag}
\alias{network_is_dag}
\title{Test if flow network is acyclic}
\usage{
network_is_dag(fl, ID = "id", toID = "toid")
}
\arguments{
\item{fl}{sf object with flowlines}

\item{ID}{character column name}

\item{toID}{character column name}
}
\value{
logical
}
\description{
Test if flow network is acyclic
}
\keyword{internal}
