% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{node_geometry}
\alias{node_geometry}
\title{Node geometry from line endpoints}
\usage{
node_geometry(x, position = "end")
}
\arguments{
\item{x}{An `sf` object (typically LINESTRING/MULTILINESTRING).}

\item{position}{Character string, either `"end"` (default) or `"start"`,
forwarded to [get_node()] to choose which node to extract.}
}
\value{
An `sf` object with geometry set to the requested node locations.
}
\description{
Replaces the geometry column of an `sf` object with point geometries taken
from each feature's line endpoints (or start nodes). Thin wrapper around
[get_node()].
}
\examples{
\dontrun{
pts <- node_geometry(flow_sf, position = "end")
}

}
\seealso{
[get_node()]
}
