% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qml.R
\name{read_qml}
\alias{read_qml}
\title{Read a QML style file}
\usage{
read_qml(qml_file)
}
\arguments{
\item{qml_file}{Path to a `.qml` file.}
}
\value{
The file contents as a single character string.
}
\description{
Read a QML style file
}
\examples{
qml <- read_qml(system.file("qml", "nexus.qml", package = "hfutils"))
substr(qml, 1, 40)
}
