% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename geometry column of sf object}
\usage{
rename_geometry(g, name)
}
\arguments{
\item{g}{sf object}

\item{name}{new geometry name}
}
\value{
sf object with renamed geometry
}
\description{
Rename geometry column of sf object
}
