% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{simplify_process}
\alias{simplify_process}
\title{Mapshaper simplification wrapper with robust system calls and validation}
\usage{
simplify_process(catchments, keep, sys, gb = 8, force = TRUE)
}
\description{
Mapshaper simplification wrapper with robust system calls and validation
}
\keyword{internal}
