% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{union_linestrings}
\alias{union_linestrings}
\title{Fast linestring union by ID}
\usage{
union_linestrings(lines, ID)
}
\arguments{
\item{lines}{An `sf` LINESTRING/MULTILINESTRING object with an attribute
column used for grouping.}

\item{ID}{A string naming the column over which to union geometries.}
}
\value{
An `sf` lines layer unioned by `ID` (column preserved). Output is
  normalized to clean LINESTRINGs via [flowpaths_to_linestrings()].
}
\description{
Fast union/merge of lines by grouping column, using `terra`'s
`aggregate()` and returning an `sf` layer. Final conversion to
clean LINESTRINGs is delegated to an internal helper
`flowpaths_to_linestrings()`.
}
\examples{
\dontrun{
out <- union_linestrings(flow_sf, "group_id")
}

}
\seealso{
[flowpaths_to_linestrings()], [union_polygons()]
}
