% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_measures}
\alias{add_measures}
\title{Add length and area measures to flowpaths/divides}
\usage{
add_measures(flowpaths, divides)
}
\arguments{
\item{flowpaths}{sf LINESTRING}

\item{divides}{sf POLYGON}
}
\value{
named list of updated flowpaths and divides
}
\description{
Add length and area measures to flowpaths/divides
}
\examples{
\dontrun{
fps <- sf::read_sf("hydrofabric.gpkg", "flowpaths")
divs <- sf::read_sf("hydrofabric.gpkg", "divides")
out <- add_measures(fps, divs)
out$flowpaths$lengthkm
}
}
