% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_properties.R
\name{get_hydroseq}
\alias{get_hydroseq}
\title{Compute and add the hydrosequence to a directed acyclic network}
\usage{
get_hydroseq(x, id = "flowpath_id", toid = "flowpath_toid")
}
\arguments{
\item{x}{A data frame (or tibble) containing at least the identifier column
given by `id` and the downstream pointer column given by `toid`.}

\item{id}{Character scalar. Column name in `topology` with unique node identifiers.
Defaults to `"flowpath_id"`.}

\item{toid}{Character scalar. Column name in `topology` with the *downstream* node
identifier for each row. Use `NA` or `0` for outlets/terminals.
Defaults to `"flowpath_toid"`.}
}
\value{
The data frame `x` with an additional `hydroseq` column representing
  the hydrosequence.
}
\description{
Compute and add the hydrosequence to a directed acyclic network
}
