% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfutils-package.R
\docType{package}
\name{hfutils-package}
\alias{hfutils}
\alias{hfutils-package}
\title{hfutils: hfutils}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Core hydrofabric access, cleaning, and summary utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lynker-spatial/hfutils}
  \item Report bugs at \url{https://github.com/lynker-spatial/hfutils}
}

}
\author{
\strong{Maintainer}: Mike Johnson \email{mjohnson@lynker.com}

Authors:
\itemize{
  \item Justin Singh-Mohudpur
}

}
\keyword{internal}
