% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lynker_spatial_auth}
\alias{lynker_spatial_auth}
\title{Authenticate with Lynker Spatial}
\usage{
lynker_spatial_auth(
  token = getOption("lynker_spatial.token"),
  ...,
  libs = c("gdal", "duckdb"),
  duckdb_con = NULL
)
}
\arguments{
\item{token}{An existing OAuth2 token. If NULL, then a new token is provisioned and returned.
If the token is expired, then}

\item{...}{Unused}

\item{libs}{Supported libraries to configure auth for.}

\item{duckdb_con}{A DuckDB DBI connection to add a bearer token secret to.}
}
\value{
The `token` argument, or a newly provisioned token
}
\description{
Authenticate with Lynker Spatial
}
\details{
lynker_spatial.auth.token
}
