% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lynker_spatial_refresh}
\alias{lynker_spatial_refresh}
\title{Refresh an existing Lynker Spatial token}
\usage{
lynker_spatial_refresh(
  token = getOption("lynker_spatial.token"),
  ...,
  client = lynker_spatial_client()
)
}
\description{
Refresh an existing Lynker Spatial token
}
\keyword{internal}
