% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lynker_spatial_token}
\alias{lynker_spatial_token}
\title{Get an OAuth2 token for Lynker Spatial}
\usage{
lynker_spatial_token(..., client = lynker_spatial_client())
}
\description{
Get an OAuth2 token for Lynker Spatial
}
\keyword{internal}
