% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{tbl_http}
\alias{tbl_http}
\title{Read DuckDB File(s) over HTTP}
\usage{
tbl_http(
  urls,
  ...,
  conn = duckdb_connection(extensions = "httpfs"),
  read_func = c("read_parquet", "read_csv", )
)
}
\arguments{
\item{urls}{1 or more URLs}

\item{...}{Unused}

\item{conn}{A DuckDB connection}

\item{read_func}{The DuckDB SQL function to call against the list of urls.
Defaults to `read_parquet`.}
}
\description{
Read DuckDB File(s) over HTTP
}
