% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfio.R
\name{write_hydrofabric}
\alias{write_hydrofabric}
\title{Write a hydrofabric GeoPackage (mixed sf + non-sf)}
\usage{
write_hydrofabric(network_list, outfile, verbose = TRUE, enforce_dm = TRUE)
}
\arguments{
\item{network_list}{named list of layers (may include `sf` and plain data.frames)}

\item{outfile}{path to `.gpkg` (".gpkg" appended if missing)}

\item{verbose}{logical, show progress via `cli`}

\item{enforce_dm}{logical, enforce `hf_dm` schema (column presence)}
}
\value{
`outfile` (invisibly)
}
\description{
Write a hydrofabric GeoPackage (mixed sf + non-sf)
}
